[IMP] account, l10n_eg_edi_eta: moved L10nEgHTTPAdapter to account
For task-3347812, we needed to create the same Adapter used in l10n_eg_edi_eta. Moving the adapter to account.tools allows us to reuse the same class in different modules.

closes odoo/odoo#126971

X-original-commit: a88027af5f5eb5d0efb4d2c42d5a179aa9bd2dc1
Related: odoo/enterprise#43448
