+++ addons/l10n_eg_edi_eta/models/account_edi_format.py
@@ -7,7 +7,7 @@ import logging
 import requests
 from werkzeug.urls import url_quote
 from base64 import b64encode
-from urllib3.util.ssl_ import create_urllib3_context
+from odoo.addons.account.tools import LegacyHTTPAdapter
 
 from odoo import api, models, _
 from odoo.tools.float_utils import json_float_round
@@ -23,21 +23,6 @@ ETA_DOMAINS = {
     'token.preproduction': 'https://id.preprod.eta.gov.eg',
     'token.production': 'https://id.eta.gov.eg',
 }
-
-
-class L10nEgHTTPAdapter(requests.adapters.HTTPAdapter):
-    """ An adapter to allow unsafe legacy renegotiation necessary to connect to
-    gravely outdated ETA production servers.
-    """
-
-    def init_poolmanager(self, *args, **kwargs):
-        # This is not defined before Python 3.12
-        # cfr. https://github.com/python/cpython/pull/93927
-        # Origin: https://github.com/openssl/openssl/commit/ef51b4b9
-        OP_LEGACY_SERVER_CONNECT = 0x04
-        context = create_urllib3_context(options=OP_LEGACY_SERVER_CONNECT)
-        kwargs["ssl_context"] = context
-        return super().init_poolmanager(*args, **kwargs)
 
 
 class AccountEdiFormat(models.Model):
@@ -61,7 +46,7 @@ class AccountEdiFormat(models.Model):
         request_url = api_domain + request_url
         try:
             session = requests.session()
-            session.mount("https://", L10nEgHTTPAdapter())
+            session.mount("https://", LegacyHTTPAdapter())
             request_response = session.request(method, request_url, data=request_data.get('body'), headers=request_data.get('header'), timeout=(5, 10))
         except (ValueError, requests.exceptions.ConnectionError, requests.exceptions.MissingSchema, requests.exceptions.Timeout, requests.exceptions.HTTPError) as ex:
             return {
